+++ src/main/resources/io/reactiverse/es4x/polyfill/global.js
@@ -102,4 +102,4 @@
     stdin: System.in
   }
 
-})(this);
+})(global || this);