+++ src/main/resources/io/reactiverse/es4x/polyfill/promise.js
@@ -227,4 +227,4 @@
   // Install (or replace) the Promise implementation
   global.Promise = Promise;
 
-})(this);
+})(global || this);