+++ src/test/java-9/io/reactiverse/es4x/dynalink/DinalynkTest.java
@@ -12,8 +12,9 @@ import org.junit.Rule;
 import org.junit.Test;
 import org.junit.runner.RunWith;
 
-import static org.junit.Assert.assertEquals;
-import static org.junit.Assert.assertTrue;
+import java.time.Instant;
+import java.util.Date;
+
 import static org.junit.Assume.assumeTrue;
 
 @RunWith(VertxUnitRunner.class)
@@ -39,6 +40,15 @@ public class DinalynkTest {
     return o.toJson().encodePrettily();
   }
 
+  public static String testInstant(Instant instant) {
+    System.out.println(instant.toString());
+    return "OK";
+  }
+
+  public static String testDate(Date instant) {
+    System.out.println(instant.toString());
+    return "OK";
+  }
   @Test(timeout = 10000)
   public void testCasting(TestContext should) throws Exception {
     final Async test = should.async();
@@ -57,6 +67,28 @@ public class DinalynkTest {
     should.assertEquals(new HttpServerOptions().toJson().encodePrettily(), loader.eval(
       "var DynalinkTest = Java.type('io.reactiverse.es4x.dynalink.DinalynkTest');\n" +
         "DynalinkTest.testDataObject({foo: 'bar'});\n"));
+
+    test.complete();
+  }
+
+  @Test(timeout = 10000)
+  public void testInstant(TestContext should) throws Exception {
+    final Async test = should.async();
+
+    should.assertEquals("OK", loader.eval(
+      "var DynalinkTest = Java.type('io.reactiverse.es4x.dynalink.DinalynkTest');\n" +
+        "DynalinkTest.testInstant(new Date());\n"));
+
+    test.complete();
+  }
+
+  @Test(timeout = 10000)
+  public void testDate(TestContext should) throws Exception {
+    final Async test = should.async();
+
+    should.assertEquals("OK", loader.eval(
+      "var DynalinkTest = Java.type('io.reactiverse.es4x.dynalink.DinalynkTest');\n" +
+        "DynalinkTest.testDate(new Date());\n"));
 
     test.complete();
   }